Pull denormalizer to CustomLandingPage module
